Fix group leaderboard crash: stats are floating-point not Int
The backend returns duration_seconds as Float (SQLite REAL) so SUM() and
MAX() produce floating-point values like 11.801... and 7.640...
Swift's JSONDecoder refuses to coerce a JSON float into Int, throwing
typeMismatch → 'data couldn't be read because it is missing'.

Change GroupLeaderboardStats.totalDuration and bestPlank from Int to
Double, and update GroupLeaderboardRow.formatDuration to accept Double.

+++ PlankChallenge/PlankChallenge/Services/API/Models/APIModels.swift
@@ -792,9 +792,12 @@ struct GroupLeaderboardEntry: Decodable, Identifiable, Sendable {
     }
     
     struct GroupLeaderboardStats: Decodable, Sendable {
-        let totalDuration: Int
+        /// Total plank duration in seconds — stored as Double because the
+        /// backend returns floating-point values from SUM(duration_seconds).
+        let totalDuration: Double
         let plankCount: Int
-        let bestPlank: Int
+        /// Best single plank in seconds — also floating-point from the DB.
+        let bestPlank: Double
     }
 }
